I deleted the object list of Logger from UI Manager because it wont work when the Logger class is static. I already have _filePath hard coded to "logs.json". So the Logger will automatically use logs.json as an file.
I added a method in the user class called "AttemptLogin" because I needed something to test the Logger class. Exceptions error messages often doesnt happen so I used a logical error for example the user logged in with wrong password. Before the method was called CheckPassword. But the new method is working like CheckPassword as usual. The only difference is that its still writes down the same error message as CheckPassword on the console. But this time its logging the error in Logger class. The return value is the same (false if the password is wrong and true when its right.)

The code in the UI Manager that controls if the password is right.. I changed that code, but it stills control the password but this time it calls the method "AttemptLogin".

In Admin I deleted all the instance related to Logger. The private readonly and the arguments Logger logger in the constructor. They will never come to use anyway. Because they wont work when the Logger class is static.