I added a few more error messages. For invalid password, emailadress and verification code.
You can see 254 is a new code I added. Its similar to the method isValidEmail on 232. But i realised when i tried to log in with wrong email and a right registered password the program crashed. We have a error message for invalid email but not for a missing user. So i added 254 so the program will not crash again. Check if its logic.

Deleted the finditem method from datastore. Because theres already a LINQ  search in UI Manager Login especially. Car class is independent and doesnt use datastore. So the question is if we should delete the list for car in UI manager because we only use dummy data.

Changed the comments and deleted some.